More fixes

+++ pkg/miles/miles.go
@@ -16,7 +16,7 @@ func DurationToMiles(dur time.Duration) float32 {
 }
 
 // ForUser returns the miles for a given user
-func ForUser(user string) int {
+func ForUser(user string) float32 {
 	evnts := []events.Event{}
 	query := fmt.Sprintf("SELECT username, event, date_created from events where username = '%s' AND event in ('login', 'logout')", user)
 	err := database.DBCon.Select(&evnts, query)
